pyproject.toml
[tool.poetry]
name = "tinyDisplay"
version = "0.1.0"
description = ""
authors = ["dhrone <ron@ritchey.org>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
pillow = "^7.2.0"
simple-pid = "^0.2.4"
cbor2 = "^5.1.2"
"luma.core" = "^1.16.0"
pytz = "^2020.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
coverage = "^5.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
